+++ commcare_connect/users/helpers.py
@@ -1,14 +1,8 @@
 import httpx
-from allauth.utils import build_absolute_uri
 from django.conf import settings
-from django.urls import reverse
-from django.utils.translation import gettext
 
-from commcare_connect.connect_id_client import send_message
-from commcare_connect.connect_id_client.models import Message
 from commcare_connect.organization.models import Organization
 from commcare_connect.utils.commcarehq_api import CommCareHQAPIException
-from commcare_connect.utils.sms import send_sms
 
 
 def get_organization_for_request(request, view_kwargs):
@@ -47,26 +41,3 @@ def create_hq_user(user, domain, api_key):
         )
 
     return hq_request.status_code == 201
-
-
-def invite_user(user, opportunity_access):
-    invite_id = opportunity_access.invite_id
-    location = reverse("users:accept_invite", args=(invite_id,))
-    url = build_absolute_uri(None, location)
-    body = (
-        "You have been invited to a new job in Commcare Connect. Click the following "
-        f"link to share your information with the project and find out more {url}"
-    )
-    if not user.phone_number:
-        return
-    send_sms(user.phone_number, body)
-    message = Message(
-        usernames=[user.username],
-        title=gettext(
-            f"You have been invited to a CommCare Connect opportunity - {opportunity_access.opportunity.name}"
-        ),
-        body=gettext(
-            f"You have been invited to a new job in Commcare Connect - {opportunity_access.opportunity.name}"
-        ),
-    )
-    send_message(message)
